Remove client side port guessing from tsh puttyconfig (#44572)
If no port was provided puttyconfig would default to 3022. That
is not guaranteed to be the correct port and causes UX issues. To
allow the port to be optional and guarantee the correct port is
used puttyconfig is updated to use 0 instead of 3022. This is a
special port that indicates to Teleport to use whatever port is
defined in the node when making dial request. Since the value
from puttyconfig is fed into tsh proxy ssh this value is honored
by Teleport during connections.