discover: generate config with --silent flag
When you add a node through the discover flow, we use the
teleport node configure command to generate a configuration file,
and then we actually start Teleport.

By default, the configure command prints some hint messages telling
the user to run teleport start, but this is not necessary in the discover
flow since we start it up for you.

The --silent command was added to solve this, but we never went back
and updated the install script to use it.

Closes #29518