Omit prerelease information from rejected connection alert
The oldest supported version used in the version check has a prerelease
of `-aa` to account for any alpha,beta,rc releases. Since the alert
blindly included said version it resulted in alerts including
`<15.0.0-aa` in the message causing confusion. This change ensures
that only the major, minor, and patch release are included in any
user facing messaging when connections are rejected.