[v16] VNet daemon: Set code signing requirements (#44745)
* Move Go error definitions to common_darwin.go

Both client_darwin.go and serivce_darwin.go are going to use them,
so it doesn't make sense to keep them in either of those files or scattered
across them.

* Roll protocol_darwin into common_darwin

There's little need for those two files to be kept separate.

* Add function for getting code signing requirement

Both the client and the daemon are going to set a requirement towards
each other, hence why this function is in common.

* Set code signing requirement towards daemon client

* Set code signing requirement towards daemon service

* Add info about mdfind to README

* Keep #cgo declarations in a single file

* Fix typos

* Move const to where it's used