Fix automatic updater installation instructions (#44164)
Closes #43865

For self-hosted Teleport Enterprise users, the current instructions show
you how to use the one-line installation script to install Teleport on a
Linux server. This script does not install the updater if the second
argument is set to `enterprise`.

Instead, instruct the user to visit the Installation page section for
package managers, install the `teleport` binary, and use the package
manager to install the updater.

An alternative approach would be to use a nested `Tabs` component with
instructions for different package managers, but nested `Tabs` look
unsightly on our docs site.