Increase the amount of time session trackers linger
When a session completes, the correspondingsession tracker is put
into a terminated state and its expiry is updated so that it will
naturally expire shortly in the future.

Teleport's upload completer runs periodically and looks for uploads
for which there is not a session tracker. If it finds any, it assumes
there was an error and that no more data will be uploaded so it
completes or aborts the upload.

There is a race condition here - if the TTL of the terminated session
tracker is shorter than the time it takes for the session to be
uploaded, then the upload completer may mistakenly determine that
a session tracker which just recently expired (and is still being
uploaded) needs to be completed, causing the upload to be cut short.

Increase the TTL of terminated sessions to drastically reduce the
chances of this race occurring. Note that the web UI's active sessions
page already filters out trackers with a state of terminated, so there
should not be any user-visible change here.