Backport part of the typography changes (#44811)
This is not a typical backport. It doesn't change the UI appearance, but
instead aims to make it a bit easier to backport changes to v16 by providing
the same additional components as v17 have. The new styles are prefixed with
the word "new", so if you are specifying typography manually, you are still
responsible for verifying that it works as intended and changing the typography
attributes if required.