Add OneLogin entity descriptor URL instructions (#44946)
Closes #24803

Show how to include the entity descriptor URL for OneLogin when
configuring the authentication connector. In `tctl sso configure`, the
flag is the same if you are using a URL or a local file.

The current guide includes a separate H3 for downloading the XML file.
This change uses a single set of numbered steps to make the instructions
easier to follow.