fix: Clarify tbot proxy-server/auth-server validation message
According to the validation logic in [`b.cfg.Address()`][1], users
should not set both the proxy-server **and** the auth-server.

English is not my first language, feel free to improve my message :).

I hope this change will help other users. It took me a while to
understand that providing both the proxy-server and the auth-server was
an error.

[1]: https://github.com/gravitational/teleport/blob/aae05c570c129f98cd1fe0588bf67dfd0804cc63/lib/tbot/config/config.go#L329-L337