buildbox: Cache git fetch directories
When fetching sources via git, first clone into the download directory
before cloning a second time from the download directory to the final
source build location. This second clone is pretty quick so does not add
much extra time in the base case, but subsequent clones are much faster
without needing the network.

This is particularly useful during development of the buildbox as the
download directory is a persistent cache, so everything only needs to be
cloned once, once and for all, rather than four times each build (once
per architecture, and on each and every build).