buildbox: Add self-built clang-12 to buildbox
Build clang-12 into the buildbox so that it can be used to build
`boringssl` for FIPS compliance. A specific version of clang is built
as specified by the instructions for a FIPS-compliant build of boringssl
(see link).

As this is now built into the buildbox, remove the apt
installation of clang-12. Install llvm-12 instead as that include other
parts of the clang toolchain not build by clang (such as `ar`).

Link: https://csrc.nist.gov/CSRC/media/projects/cryptographic-module-validation-program/documents/security-policies/140sp4407.pdf