buildbox: Add clang 12.0.0 to buildbox
Add clang 12.0.0 to the buildbox in `/opt/clang` so that the `boring-rs`
crate can build `boringssl` for the FIPS build of Teleport. This
specific version of clang is required for FIPS compliance.

To make this work we need to ensure some arguments are always passed to
clang (`--gcc-toolchain` and `--sysroot`) so we also add a front-end
shell script that allows extra arguments to be passed to clang via the
environment.

Also add a few symlinks as it seems cmake that builds boringssl needs
some clang/llvm tools to be specifically named and located:

    /opt/clang/bin/clang++-12
    /usr/bin/llvm-ar-12
    /usr/bin/llvm-ranlib-12