[sec_scan][27] add support for LDAP users and macOS (#45109)
* [sec_scan][27] add support for LDAP users and macOS

This PR extends support for authorized keys report for users managed by LDAP system and macOS targets.

It leverages `getpwent` to read the system database files and retrieve the user properties.
It doesn't use the `getpwent_r` because it's not available in macOS and because it's not (yet) standerdized

>   PLEASE NOTE: the `getpwent_r' function is not (yet) standardized.
>   The interface may change in later versions of this library.  But
>   the interface is designed following the principals used for the
>   other reentrant functions so the chances are good this is what the
>   POSIX people would choose.

Part of https://github.com/gravitational/access-graph/issues/637

* handle comments

* handle comments 2

* add comment