WIP: buildbox: Build boringssl in buildbox
Build boringssl with clang to be used with the FIPS build of teleport.
boringssl can be built by the boring-rs crate but because we are cross
compiling, having the rust build script of boring-rs invoke cmake /
clang / ninja correctly for cross compiling can be tricky. Instead we
may be able to build boringssl first with greater control and tell the
boring-rs crate to use the pre-build boringssl build.

This does not currently work. The current recipes in the buildbox will
download the correct version of boringssl and set up a basic toolchain
file. It does not build it because I cannot get the correct magic
incantation to make it work.

This commit exists as a partial step and it can be used interactively to
experiment with the build:

    make -C build.assets buildbox-thirdparty-libs
    # wait quite a while (> 1 hour)
    docker run -it --rm buildbox-thirdparty-libs:teleport17
    ctr$ cd /opt/thirdparty/amd64/src/boringssl<tab>
    ctr$ # edit build/toolchain
    ctr$ cmake -B build -G Ninja -DFIPS=1 \
      -DCMAKE_TOOLCHAIN_FILE=build/toolchain \
      -DCMAKE_BUILD_TYPE=RELEASE \
      .
    ctr$ ninja -C build # if you get this far