Fix OpenSSH guide sidebar highlighting
Backports #45236

Currently, the agentless OpenSSH guide does not appear on the sidebar.
This is because the sidebar generator expects menu pages to be named
after their corresponding subdirectory and to exist at either (a) the
same level as the subdirectory or (b) within the subdirectory. Since a
page exists at both locations called `openssh.mdx`, the sidebar
generator only shows one of these.

This change edits the OpenSSH guide to be called `openssh-agentless.mdx`
so it appears on the sidebar.