Restore enterprise auth connectors in tctl get all output
When enterprise tctl was removed in v16 the ability to retrieve
SAML and OIDC connectors via tctl get all was inadvertently removed.
This ensures a resource marshaler and unmarshaler is registered for
the two types so that the tctl get all machinery includes them in
the group of resources retrieved. Tests were updated to ensure that
tctl get all output contains the expected resource types.