Prevent exiting a session prior to output being consumed (#45223)
#17687 attempted to fix flakiness of TestIntegrations/AuditOn by
sending an exit-status request _prior_ to consuming all output from
the PTY. While this made the test more reliable, it created a
scenario that allowed for a session to be completed without all of
the data from the PTY being consumed by the client. This condition
was hit by running an ansible playbook that output 1MB to stdout.

The reason TestIntegrations/AuditOn was flaky is because the
exit-status request was not received at times. The mechanism used
to send that request requires sending the result over a channel
and the request to be sent by another goroutine. That provides an
opportunity for the request on the channel to be processed after
the underlying ssh connection has been closed.

To resolve the issue of missing output, the change in order of
operations from #17687 was reverted and the exit-status request
is now being sent directly in the same goroutine that waits
for the session to end instead. This change now causes the exit-status
to be sent later in time, which in the real world should not be
noticed, however, some time dependent tests needed to have their
timeout for sessions completing bumped.