Deprecate ListResources helpers and migrate to the unified API
While the ListUnifiedResources API was first introduced to support
a redesigned Web UI, it has evolved into a more performant and
less resource hungry API than the ListResources API. Most of the
client tools retrieval of ssh servers has already been converted
to use the unified API to reduce resource consumption of Auth and
to improve the amount of time it takes to retrieve results, especially
in clusters with a large number of servers.

This migrates most of tctl and tsh invocations that looked up resources
to use the unified API. Flows that are specific to the Web UI and
Connect have been left on the legacy API for one reason: resource
counts. The unified API does not return a TotalCount, and there
are a few views in both the Web UI and Connect that consume the
count. In an effort not to make the UX worse, those cases have been
left to use the legacy API and should be converted in the future
when the frontend no longer displays a count.

To accomodate these use cases that have not been converted the
existing ListResources helpers have been left in place. However,
the helpers have been deprecated to signal to any new code that
they should use the unified API instead.