Fix the Okta SSO guide (#45346)
It looks like the guide intends for you to set up two Okta groups
(okta-dev and okta-admin) that map to Teleport roles.

The Teleport roles that the guide refers to are incorrect though.
For the okta-dev group, the guide has to create a "dev" role, but
then configures an SSO connector to assign you the preset "access"
role and the "dev" role goes unused.

Additionally, the okta-admin group refers to an "admin" role, but the
SSO connector that you set up maps users to the "editor" role.

Correct the guide to properly use the "dev" role that it asks you to
create and to replace mentions of "admin" with "editor".

Closes #28660