Fix prose linting issues in the changelog (#45527)
Backports #44969

The changelog contains a number of inconsistencies with the rest of the
Teleport documentation. For example, while we have moved away from using
"Application Access", "Database Access" and so on in favor of describing
Teleport Access in a more integrated way, the changelog still uses the
outdated phrasing throughout. The changelog also fails to capitalize
Teleport capabilities such as Access Requests and Device Trust.

This change fixes prose linter errors in the changelog to make it more
consistent with the rest of the docs.