Fixes App-deleting issue in Okta Sync (#4874)
* Fixes App-deleting issue in Okta Sync

The Okta Sync service treats a connectivity failure to the Okta
organization when fetching application data as if the queried application
were missing, and continues on with the Sync process. This results in
Teleport mistakenly deleteing the target Okta app.

This patch makes the Sync Service treat all errors when querying an
appliction (other than Not Found) a showstopper, and abort the current
sync. Hopefully, by the time the next sync rolls around the network will
be back up and the next sync will carry on as normal.

Changelog: Fixed issue in Okta Sync that spuriously deletes Okta Applications due to connectivity errors.

* linter

Source-commit: gravitational/teleport.e@5816524337ba75e4212415eb4dcd3d6f33cab2a4
Source-pull-requests: gravitational/teleport.e#4874