web: be more explicit about root/leaf clusters (#45531)
We use the terms "root" and "leaf" pervasively in our docs and
our conversations with customers, but the web UI tends to refer
to everything as a "trusted cluster" without distinguishing
between root and leaf. After noticing us incorrectly describing
which clusters are roots and which are leafs, I decided we should
make the UI self-explanatory.

- change sidebar title to indicate that the Trusted Clusters page
  is listing trusted *root* clusters
- update the manage clusters page to label leaf clusters in addition
  to roots