build: Update Makefiles to enable FIPS build with buildbox-ng
Update the Makefiles so that we can use buildbox-ng to make a FIPS
release build of Teleport.

This is done by adding a new release target `release-ng-amd64-fips` that
builds just the enterprise edition of Teleport with FIPS enabled and
updating the top-level `Makefile` to set some environment variables
needed for a FIPS build and a target to build just an enterprise
release.