desktop access: improve error handling when directory sharing fails
If drive redirection is disabled on the Windows side, then directory
sharing will fail even if Teleport RBAC allows it. Prior to this change
failure to share a directory was considered a fatal error that would
terminate the session.

This change updates the error message and makes the error a warning
so that the session remains alive even if the directory sharing
operation fails.