build: Add support for FIPS builds with buildbox-ng (#4915)
Tweak the build so that we can do a FIPS build of Teleport with
buildbox-ng (the cross-compiling buildbox).

This updates the `teleport-fips-check` target to use `$(BUILDFLAGS)`
when doing the test build of teleport to check if it has the
boringcrypto symbols in it. The build now requires `-buildmode=pie` as
the C libraries in the buildbox are compiled as PIE (position
independent executable) libraries. The subsequent `-ldflags ""`
overrides the `-w -s` from `$(BUILDFLAGS)` as those flags remove the
symbol table(s) that are needed to check if the binary was correctly
built with boringcrypto.

We also need to build rdpclient with some extra environment variables,
via the `$(RDPCLIENT_ENV)` variable set in the OSS Makefile. This sets
`BORING_BSSL_FIPS_SYSROOT` to point to the cross-compiling toolchain and
`CGO_ENABLED=0` as the build of `boring-sys` builds and runs a Go binary
as part of its build and there are weird cross-compiler issues with that
if CGO_ENABLED is on (and it does not seem to be needed to be on).

On a non-cross build, `$(RDPCLIENT_ENV)` is not set so there is no
change to the build in the default case.

Source-commit: gravitational/teleport.e@9d131e3746ca57a56eec755c026b034d7df9421d
Source-pull-requests: gravitational/teleport.e#4915