Add ownership tags to AWS OIDC EKS Enrollment (#44736) (#45725)
* Add ownership tags to AWS OIDC EKS Enrollment

This PR adds the ownership tags to the temporary EKS Access Entry
created during the EKS Enrollment.

There might happen that the user already completed the set up once, and
so they won't be asked to run the configuration script (using oneoff +
cloudshell) again. If there's a access denied, the Access Entry is
created without the tags.

With or without the tags, the Access Entry is removed after the
enrollment process.

* log message asking users to add eks:TagResource action

* fix indentation in docs

* fix new tests