[v16] Fix AWS App Access creation for AWS OIDC Integration when using the account number as name (#45819)
* Fix AWS App Access creation for AWS OIDC Integration

If the integration name is digits only, the resulting address for the
application will look like this:
`123456789012.proxy.example.com:443`
This fails to parse with go's `url.Parse`.

This PR keeps the existing logic for creating the AWS App Access but
does a best effort to fix this issue by removing the `:443` from the
public proxy addr.
If another port is used, we would still get the error.

* prepend the protocol so that url.parse accepts any port number

* move change to types/app