Add proxied databases to DatabaseService
This PR changes the DatabaseService to also heartbeat the proxied
databases.
This is useful to troubleshoot issues with database access.

This was done in a way that we can extend the properties of each
database.
Eg, in the future we can also include the number of active connections
to a given database.