Add favicon theme listener (#45919)
This PR changes the way we update our favicon based on system theme
preference. Currently, we rely on the media attr in the link tag to obey
the system preference, but some browsers don't seem to support that very
well. Instead, we will now add a listener when Teleport loads to the
theme and set the favicon dynamically.

Note: this means the favicon will match the theme of the system/browser
and NOT the theme of the page (which can be different, depending on
user). The goal of the favicon theme switch is to make the favicon
visible based on the browser's theme, rather than necessarily matching
the page theme.

This can eventually be expanded upon to be included in a theme switch
when we have a user settings page that allows "system" as a theme choice