Prevent interactive sessions from hanging on exit (#45879) (#45954)
The TTY allocated for interactive sessions was not being handled
properly. The parent process was never closing its half of the
TTY, which prevented the PTY read from unblocking when the child
process terminated. This became more noticeable after #45223
landed because interactive sessions were now properly waiting for
the PTY copying to be completed before exiting.

Additionally, the PTY is no longer being provided to the child reexec
process since it should be only be used by the parent process.

Closes #45878.