[v16] build: Update buildbox-ng for FIPS builds
* buildbox: Cache git fetch directories

When fetching sources via git, first clone into the download directory
before cloning a second time from the download directory to the final
source build location. This second clone is pretty quick so does not add
much extra time in the base case, but subsequent clones are much faster
without needing the network.

This is particularly useful during development of the buildbox as the
download directory is a persistent cache, so everything only needs to be
cloned once, once and for all, rather than four times each build (once
per architecture, and on each and every build).

* buildbox: Add clang 12.0.0 to buildbox

Add clang 12.0.0 to the buildbox in `/opt/clang` so that the `boring-rs`
crate can build `boringssl` for the FIPS build of Teleport. This
specific version of clang is required for FIPS compliance.

To make this work we need to ensure some arguments are always passed to
clang (`--gcc-toolchain` and `--sysroot`) so we also add a front-end
shell script that allows extra arguments to be passed to clang via the
environment.

Also add a few symlinks as it seems cmake that builds boringssl needs
some clang/llvm tools to be specifically named and located:

    /opt/clang/bin/clang++-12
    /usr/bin/llvm-ar-12
    /usr/bin/llvm-ranlib-12

* build: Update Makefiles to enable FIPS build with buildbox-ng

Update the Makefiles so that we can use buildbox-ng to make a FIPS
release build of Teleport.

This is done by adding a new release target `release-ng-amd64-fips` that
builds just the enterprise edition of Teleport with FIPS enabled and
updating the top-level `Makefile` to set some environment variables
needed for a FIPS build and a target to build just an enterprise
release.