WIP Remove RootRef
There are still failing Popover and Modal tests.

In general, RootRef was for finding the DOM node of its children. Such
use case is hard to support without findDOMNode. We could consider using
a wrapper div [1], but I worry this might break too many callsites.

Instead, I'm trying to pass a ref, though this will not work if the child
is a React fragment.

Since we render a wrapper anyway, wouldn't it be better to put a ref on
the wrapper and get to the actual child node this way?

[1]: https://react.dev/reference/react-dom/findDOMNode#adding-a-wrapper-div-element