build: Remove obsolete cbindgen build in rdpclient target (#4969)
Remove the build of `cbindgen` in the `rdpclient` target of the Makefile
as it is built by `lib/srv/desktop/rdp/rdpclient/build.rs`. This change
was made in gravitational/teleport#14177 in the OSS Makefile but the
equivalent change was not made in the Enterprise Makefile.

Link: https://github.com/gravitational/teleport/pull/14177

Source-commit: gravitational/teleport.e@487320ad17e3a33d5068ebd623a3593739c47e8d
Source-pull-requests: gravitational/teleport.e#4969