Make the Cloud migration guide more general (#46074)
Closes #3683

We do not have a guide for moving from Community Edition to Enterprise.
The steps are nearly identical to those presented in the guide to
migrating from Enterprise (Self-Hosted) to Enterprise (Cloud). This
change repurposes the Cloud migration guide so it applies to migrations
between Teleport plans in general.

This change also moves the migration guide to the root of Admin Guides,
since it's a procedure that we want to make prominent as we encourage
OSS users to move to Enterprise.