Fix removing nodes with bulkAction remove (#4981)
We recently made a fix (again) to properly show hostnames in the request
checkout (...again) but we weren't properly removing this item from the
request checkout. This would mean if someone clicked the bulk
"add/remove from request" button and a node was selected, it would just
stay in the checkout (the user would still see this, so its just an
annoyance).

Also, all our tests around this only checked _adding_ nodes so I added a
test to check _removing_ nodes as well.

requires https://github.com/gravitational/teleport/pull/45964 for a
`testid` to be in place

Source-commit: gravitational/teleport.e@82dbe2e3b8336285da24859c86922c1a331f8027
Source-pull-requests: gravitational/teleport.e#4981