[v16] Get PROXY protocol guide to match the style guide (#46312)
* Get PROXY protocol guide to match the style guide

Closes #32168

Format the guide as a how-to guide (rather than a reference or
architectural guide). This is appropriate, since the guide is a detailed
look at a single configuration option that requires some architectural
planning for users can enable it.

Also add an introductory section that describes the scope of the guide,
and make various edits for readability and consistency with the rest of
the docs site.

* Respond to feedback

- Include "PROXY Protocol" in the title.
- Edit the description.