Preserve integration roles in SAML role calculation (#4995)
* Preserve integration roles in SAML role calculation

The SAML role calculation introduced in #4849 clobbers any roles outside
of those calculated by the SAML connector trait-to-role mapping. This
breaks the Okta integration, which requires the `okta-requester` role
to work.

This patch addresses the above issue by introducing the idea of
"preserved roles" for a given origin. Any roles in the "preserved" list
for a given user's origin will be preserved through the role calcualtion
process.

* test fixups

Source-commit: gravitational/teleport.e@8a5e69c185da82816d442a5fa0d0a9c16d2b2ac3
Source-pull-requests: gravitational/teleport.e#4995