Remove findDOMNode usage from Modal and Popover (#46123)
* Remove disablePortal prop

It was not used within the app anyway

* Get rid of ownerDocument and ownerWindow

The only situation where they'd be needed if we were handling some nodes
between iframes, but we don't use <iframe> even once in our code.
Calls to those functions can be replaced with document and window.

https://stackoverflow.com/questions/9845043/when-node-ownerdocument-is-not-window-document
https://stackoverflow.com/questions/9183555/whats-the-point-of-document-defaultview

* Remove container prop for Modal, replace Portal with createPortal

Initially I wanted to just remove the container prop, but as I sat down
to it, I realized that the whole Portal component is not necessary as
we can just replace it with calling createPortal directly.

The onRendered prop of Portal was never used, rendering the handleRendered
function unnecessary.

Similarly, handlePortalRef was set to the ref prop on Portal. This
function sets mountNode, but after we replaced ownerDocument with document,
it's mountNode is no longer being used.

* Refactor Modal to use TypeScript

* Remove RootRef

dialogRef used to point at the child of StyledModal. To get rid of
findDOMNode, we replaced it by taking children directly from the ref
set on StyledModal.

* Remove findDOMNode from Popover

* Fix types for mouse & keyboard events

* Insert whitespace between prop type declarations

* Remove link from BackdropProps comment

* Extract type for BackdropProps

* Add JSDoc for children prop