Update generic storage service to support validation and custom keys (#46319)
Adds two new optional configuration options to the generic.Service:
ValidateFunc and KeyFunc.

ValidateFunc is a custom function that is called prior to persisting
a resource in the backend. If the function returns an error, the
resource is not stored and the error is returned to users.

KeyFunc is a custom function used to derive the key for a particular
resource. By default the generic service uses the metadata name as
the key, however, in some scenarios more control over the key might
be desired. For instance, a singleton resource might want to enforce
that the key is also a static value instead of something that a
user may supply.