[v16] Adds a detailed status block to Okta plugin resource (#46387)
* [v16] Adds AppName to Okta Plugin SyncSettings

Backports #45906

Recording the App Name is useful for generating links on the Okta status page,
so a field to record it has been added to the Okta plugin Sync Settings
record.

* [v16] Adds detailed Okta status to Plugin Resource

Backports #44431

Once we collect the Okta usage statistics we need a durable place to store
them. This patch adds a `Details` block to the Okta plugin status in order
provide such a location. As an extra bonus, we already know that the plugin
status can be reached from both the running plugin and the UI code making it
easier plumb the data through in a subsequent PR.