[aws_sync] reuse existing resources on AWS failures (#46333)
Currently, if a AWS call fails due to a rate limited API or a network issue, Discovery service considers the underlying ret psources as deleted from AWS.
This causes several problems because those resources will be removed from access graph view.

This PR changes that behavior and ensures resources are reused on failures.