Migrate database server heartbeats to inventory control stream (#46118)
Adds support for database server heartbeats to the auth server and
converts agents to use the new mechanism if they detect that auth
supports it. Additionally, clean up of database servers is now
performed automatically by auth in response to a GoodBye message.

The srv.HeartBeatI interface was also updated to remove ForceSend
since nothing consumed it after the conversion. ForceSend only
existed for tests and was a source of flakiness if heartbeats were
unable to be sent before the timeout.

Updates #40917