Use the new -diff flag for the go mod tidy check (#46459)
* Use the new -diff flag for the go mod tidy check

Go 1.23 introduces a `go mod tidy -diff` option that prints the
diff and exits with a non-zero status code if updates are needed,
which eliminates the need for us to check the git state to determine
if the modules needed tidying.

* Set GOFLAGS

Disable VSC stamping so that the workflows don't depend on git.