Fix Auto Discover EC2 test when running in darwin
This PR changes the method used for calculating the checksum for the
`/etc/teleport.yaml` file.

It was using a method that doesn't allow symlinks and we can't run tests
that use it in darwin boxes.

We don't expect any real issues with this, because files are in
root-only allowed directories (`/etc/`).