update default signature_algorithm_suite for new clusters
This PR updates the default signature_algorithm_suite for new v17
clusters. The new default will be balanced-v1, or hsm-v1 if an HSM is
enabled, or fips-v1 if FIPS mode is enabled (the logic is probably
best described in the code).

Existing clusters upgraded to v17 will continue to have an unspecified
suite in their cluster_auth_preference, and this will continue to be
treated as the "legacy" suite. Only newly created v17 clusters will get
one of the new suites without manually configuring it.

changelog: Elliptic curve signature algorithms will be used by default for all new v17 clusters