Fix `roleToRecipients` descriptions
Closes #45287

Access Request plugin guides describe the value of the
`roleToRecipients` field in Access Request plugin Helm charts as taking
either a single string or an array of strings as the value of each
mapping. This change fixes the error in all Access Request plugin guides
that include this Helm chart value.